Refactor giant store into modular slices
